temp
temp

+++ src/Basic1/Main.java
@@ -1,8 +1,9 @@
 package Basic1;
+//найдите значение функции z=((a-3)*b/2)+c
 
 public class Main {
     public static void main(String[] args) {
-        float z;
+        int z;
         int a=5,b=4,c=3;
         z= (a-3)*b/2 + c;
         System.out.print(z);
